feat: finalized metho to synchronized ordes in system
-Added enum to method payment
-added new endpoint in creposicion Controller
-added feigclient to new endpojt
-update service

+++ common-module/src/main/java/com/cumpleanos/core/models/entities/ReposicionPago.java
@@ -40,7 +40,7 @@ public class ReposicionPago {
     private BigDecimal monto;
 
     @Column(name = "RPA_LOTE", precision = 17, scale = 4)
-    private BigDecimal lote;
+    private Integer lote;
 
     @Column(name = "RPA_BANCO", length = 20)
     private String banco;
